% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisquare1.R
\name{chisquare1}
\alias{chisquare1}
\title{Chi Square statistics - One Way Table}
\usage{
chisquare1(obs, exp, x2 = 0)
}
\description{
Returns a list with the the chi-square test statistic, right-sided
p-value and degrees of freedom for a given one-way table.
}
\details{
Input is two vectors of observed and expected values.
}
