% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisquare2.R
\name{chisquare2}
\alias{chisquare2}
\title{Chi Square statistics - Two Wqy Table}
\usage{
chisquare2(m, x2 = 0)
}
\description{
Returns a list with the the chi-square test statistic, right-sided
p-value and degrees of freedom for a given matrix.
}
\details{
Input is a matrix of values.
}
